refactor(web): renamed google-cloud and form utils
