+++ apps/web/src/utils/google-cloud-utils.ts
